Update echodata module to reflect framework
