+++ echopype/tests/echodata/test_echodata.py
@@ -1,197 +1,19 @@
 from textwrap import dedent
-
-import fsspec
 from pathlib import Path
 import shutil
 
 from datatree import DataTree
 from zarr.errors import GroupNotFoundError
 
-import echopype
-from echopype.calibrate.env_params import EnvParams
-from echopype.echodata import EchoData
-from echopype import open_converted
+from echopype.echodata.echodata import EchoData
 
 import pytest
 import xarray as xr
 import numpy as np
 
-from echopype.testing import check_consolidated
-
-
-@pytest.fixture(scope="module")
-def single_ek60_zarr(test_path):
-    return (
-        test_path['EK60'] / "ncei-wcsd" / "Summer2017-D20170615-T190214__NEW.zarr"
-    )
-
-
-@pytest.fixture(
-    params=[
-        single_ek60_zarr,
-        (str, "ncei-wcsd", "Summer2017-D20170615-T190214.zarr"),
-        (None, "ncei-wcsd", "Summer2017-D20170615-T190214__NEW.nc"),
-        "s3://data/ek60/ncei-wcsd/Summer2017-D20170615-T190214.nc",
-        "http://localhost:8080/data/ek60/ncei-wcsd/Summer2017-D20170615-T190214.zarr",
-        "s3://data/ek60/ncei-wcsd/Summer2017-D20170615-T190214.zarr",
-        fsspec.get_mapper(
-            "s3://data/ek60/ncei-wcsd/Summer2017-D20170615-T190214.zarr",
-            **dict(
-                client_kwargs=dict(endpoint_url="http://localhost:9000/"),
-                key="minioadmin",
-                secret="minioadmin",
-            ),
-        ),
-    ],
-    ids=[
-        "ek60_zarr_path",
-        "ek60_zarr_path_string",
-        "ek60_netcdf_path",
-        "ek60_netcdf_s3_string",
-        "ek60_zarr_http_string",
-        "ek60_zarr_s3_string",
-        "ek60_zarr_s3_FSMap",
-    ],
+from echopype.testing import (
+    _check_consolidated,
 )
-def ek60_converted_zarr(request, test_path):
-    if isinstance(request.param, tuple):
-        desired_type, *paths = request.param
-        if desired_type is not None:
-            return desired_type(test_path['EK60'].joinpath(*paths))
-        else:
-            return test_path['EK60'].joinpath(*paths)
-    else:
-        return request.param
-
-
-@pytest.fixture(
-    params=[
-        (
-            ("EK60", "ncei-wcsd", "Summer2017-D20170615-T190214.raw"),
-            "EK60",
-            None,
-            None,
-            "CW",
-            "power",
-        ),
-        (
-            ("EK80_NEW", "D20211004-T233354.raw"),
-            "EK80",
-            None,
-            None,
-            "CW",
-            "power",
-        ),
-        (
-            ("EK80_NEW", "echopype-test-D20211004-T235930.raw"),
-            "EK80",
-            None,
-            None,
-            "BB",
-            "complex",
-        ),
-        (
-            ("EK80_NEW", "D20211004-T233115.raw"),
-            "EK80",
-            None,
-            None,
-            "CW",
-            "complex",
-        ),
-        (
-            ("ES70", "D20151202-T020259.raw"),
-            "ES70",
-            None,
-            None,
-            None,
-            None,
-        ),
-        (
-            ("AZFP", "ooi", "17032923.01A"),
-            "AZFP",
-            ("AZFP", "ooi", "17032922.XML"),
-            "Sv",
-            None,
-            None,
-        ),
-        (
-            ("AZFP", "ooi", "17032923.01A"),
-            "AZFP",
-            ("AZFP", "ooi", "17032922.XML"),
-            "TS",
-            None,
-            None,
-        ),
-        (
-            ("AD2CP", "raw", "090", "rawtest.090.00001.ad2cp"),
-            "AD2CP",
-            None,
-            None,
-            None,
-            None,
-        ),
-    ],
-    ids=[
-        "ek60_cw_power",
-        "ek80_cw_power",
-        "ek80_bb_complex",
-        "ek80_cw_complex",
-        "es70",
-        "azfp_sv",
-        "azfp_sp",
-        "ad2cp",
-    ],
-)
-def compute_range_samples(request, test_path):
-    (
-        filepath,
-        sonar_model,
-        azfp_xml_path,
-        azfp_cal_type,
-        ek_waveform_mode,
-        ek_encode_mode,
-    ) = request.param
-    if sonar_model.lower() == 'es70':
-        pytest.xfail(
-            reason="Not supported at the moment",
-        )
-    path_model, *paths = filepath
-    filepath = test_path[path_model].joinpath(*paths)
-
-    if azfp_xml_path is not None:
-        path_model, *paths = azfp_xml_path
-        azfp_xml_path = test_path[path_model].joinpath(*paths)
-    return (
-        filepath,
-        sonar_model,
-        azfp_xml_path,
-        azfp_cal_type,
-        ek_waveform_mode,
-        ek_encode_mode,
-    )
-
-
-@pytest.fixture(
-    params=[
-        {
-            "path_model": "EK60",
-            "raw_path": "Winter2017-D20170115-T150122.raw",
-        },
-        {
-            "path_model": "EK80",
-            "raw_path": "D20170912-T234910.raw",
-        },
-    ],
-    ids=[
-        "ek60_winter2017",
-        "ek80_summer2017",
-    ],
-)
-def range_check_files(request, test_path):
-    return (
-        request.param["path_model"],
-        test_path[request.param["path_model"]].joinpath(request.param['raw_path'])
-    )
 
 
 class TestEchoData:
@@ -295,9 +117,16 @@ class TestEchoData:
 
     def test_setitem(self, converted_zarr):
         ed = self.create_ed(converted_raw_path=converted_zarr)
-        ed['Sonar/Beam_group1'] = ed['Sonar/Beam_group1'].rename({'beam': 'beam_newname'})
+        ed['Sonar/Beam_group1'] = ed['Sonar/Beam_group1'].rename(
+            {'beam': 'beam_newname'}
+        )
 
-        assert sorted(ed['Sonar/Beam_group1'].dims.keys()) == ['beam_newname', 'channel', 'ping_time', 'range_sample']
+        assert sorted(ed['Sonar/Beam_group1'].dims.keys()) == [
+            'beam_newname',
+            'channel',
+            'ping_time',
+            'range_sample',
+        ]
 
         try:
             ed['SomeRandomGroup'] = 'Testing value'
@@ -315,15 +144,71 @@ class TestEchoData:
         assert result is None
         assert isinstance(ed_result, xr.Dataset)
 
+    @staticmethod
+    def test__harmonize_env_param_time():
+        # Scalar
+        p = 10.05
+        assert EchoData._harmonize_env_param_time(p=p) == 10.05
+
+        # time1 length=1, should return length=1 numpy array
+        p = xr.DataArray(
+            data=[1],
+            coords={
+                "time1": np.array(
+                    ["2017-06-20T01:00:00"], dtype="datetime64[ns]"
+                )
+            },
+            dims=["time1"],
+        )
+        assert EchoData._harmonize_env_param_time(p=p) == 1
+
+        # time1 length>1, interpolate to tareget ping_time
+        p = xr.DataArray(
+            data=np.array([0, 1]),
+            coords={
+                "time1": np.arange(
+                    "2017-06-20T01:00:00",
+                    "2017-06-20T01:00:31",
+                    np.timedelta64(30, "s"),
+                    dtype="datetime64[ns]",
+                )
+            },
+            dims=["time1"],
+        )
+        # ping_time target is identical to time1
+        ping_time_target = p["time1"].rename({"time1": "ping_time"})
+        p_new = EchoData._harmonize_env_param_time(
+            p=p, ping_time=ping_time_target
+        )
+        assert (p_new["ping_time"] == ping_time_target).all()
+        assert (p_new.data == p.data).all()
+        # ping_time target requires actual interpolation
+        ping_time_target = xr.DataArray(
+            data=[1],
+            coords={
+                "ping_time": np.array(
+                    ["2017-06-20T01:00:15"], dtype="datetime64[ns]"
+                )
+            },
+            dims=["ping_time"],
+        )
+        p_new = EchoData._harmonize_env_param_time(
+            p=p, ping_time=ping_time_target["ping_time"]
+        )
+        assert p_new["ping_time"] == ping_time_target["ping_time"]
+        assert p_new.data == 0.5
+
     @pytest.mark.parametrize("consolidated", [True, False])
     def test_to_zarr_consolidated(self, mock_echodata, consolidated):
         """
         Tests to_zarr consolidation. Currently, this test uses a mock EchoData object that only
-        has attributes. The consolidated flag provided will be used in every to_zarr call (which 
-        is used to write each EchoData group to zarr_path). 
+        has attributes. The consolidated flag provided will be used in every to_zarr call (which
+        is used to write each EchoData group to zarr_path).
         """
         zarr_path = Path('test.zarr')
-        mock_echodata.to_zarr(str(zarr_path), consolidated=consolidated, overwrite=True)
+        mock_echodata.to_zarr(
+            str(zarr_path), consolidated=consolidated, overwrite=True
+        )
 
         check = True if consolidated else False
         zmeta_path = zarr_path / ".zmetadata"
@@ -331,228 +216,7 @@ class TestEchoData:
         assert zmeta_path.exists() is check
 
         if check is True:
-            check_consolidated(mock_echodata, zmeta_path)
+            _check_consolidated(mock_echodata, zmeta_path)
 
         # clean up the zarr file
         shutil.rmtree(zarr_path)
-
-
-def test_open_converted(ek60_converted_zarr, minio_bucket):  # noqa
-    def _check_path(zarr_path):
-        storage_options = {}
-        if zarr_path.startswith("s3://"):
-            storage_options = dict(
-                client_kwargs=dict(endpoint_url="http://localhost:9000/"),
-                key="minioadmin",
-                secret="minioadmin",
-            )
-        return storage_options
-
-    storage_options = {}
-    if not isinstance(ek60_converted_zarr, fsspec.FSMap):
-        storage_options = _check_path(str(ek60_converted_zarr))
-
-    try:
-        ed = open_converted(
-            ek60_converted_zarr, storage_options=storage_options
-        )
-        assert isinstance(ed, EchoData) is True
-    except Exception as e:
-        if (
-            isinstance(ek60_converted_zarr, str)
-            and ek60_converted_zarr.startswith("s3://")
-            and ek60_converted_zarr.endswith(".nc")
-        ):
-            assert isinstance(e, ValueError) is True
-
-
-def test_compute_range(compute_range_samples):
-    (
-        filepath,
-        sonar_model,
-        azfp_xml_path,
-        azfp_cal_type,
-        ek_waveform_mode,
-        ek_encode_mode,
-    ) = compute_range_samples
-    ed = echopype.open_raw(filepath, sonar_model, azfp_xml_path)
-    rng = np.random.default_rng(0)
-    stationary_env_params = EnvParams(
-        xr.Dataset(
-            data_vars={
-                "pressure": ("time3", np.arange(50)),
-                "salinity": ("time3", np.arange(50)),
-                "temperature": ("time3", np.arange(50)),
-            },
-            coords={
-                "time3": np.arange("2017-06-20T01:00", "2017-06-20T01:25", np.timedelta64(30, "s"), dtype="datetime64[ns]")
-            }
-        ),
-        data_kind="stationary"
-    )
-    if "time3" in ed["Platform"] and sonar_model != "AD2CP":
-        ed.compute_range(stationary_env_params, azfp_cal_type, ek_waveform_mode)
-    else:
-        try:
-            ed.compute_range(stationary_env_params, ek_waveform_mode="CW", azfp_cal_type="Sv")
-        except ValueError:
-            pass
-        else:
-            raise AssertionError
-
-    mobile_env_params = EnvParams(
-        xr.Dataset(
-            data_vars={
-                "pressure": ("time", np.arange(100)),
-                "salinity": ("time", np.arange(100)),
-                "temperature": ("time", np.arange(100)),
-            },
-            coords={
-                "latitude": ("time", rng.random(size=100) + 44),
-                "longitude": ("time", rng.random(size=100) - 125),
-            }
-        ),
-        data_kind="mobile"
-    )
-    if "latitude" in ed["Platform"] and "longitude" in ed["Platform"] and sonar_model != "AD2CP" and not np.isnan(ed["Platform"]["time1"]).all():
-        ed.compute_range(mobile_env_params, azfp_cal_type, ek_waveform_mode)
-    else:
-        try:
-            ed.compute_range(mobile_env_params, ek_waveform_mode="CW", azfp_cal_type="Sv")
-        except ValueError:
-            pass
-        else:
-            raise AssertionError
-
-    env_params = {"sound_speed": 343}
-    if sonar_model == "AD2CP":
-        try:
-            ed.compute_range(
-                env_params, ek_waveform_mode="CW", azfp_cal_type="Sv"
-            )
-        except ValueError:
-            pass  # AD2CP is not currently supported in ed.compute_range
-        else:
-            raise AssertionError
-    else:
-        echo_range = ed.compute_range(
-            env_params,
-            azfp_cal_type,
-            ek_waveform_mode,
-        )
-        assert isinstance(echo_range, xr.DataArray)
-
-
-def test_nan_range_entries(range_check_files):
-    sonar_model, ek_file = range_check_files
-    echodata = echopype.open_raw(ek_file, sonar_model=sonar_model)
-    if sonar_model == "EK80":
-        ds_Sv = echopype.calibrate.compute_Sv(echodata, waveform_mode='BB', encode_mode='complex')
-        range_output = echodata.compute_range(env_params=[], ek_waveform_mode='BB')
-        nan_locs_backscatter_r = ~echodata["Sonar/Beam_group1"].backscatter_r.isel(beam=0).drop("beam").isnull()
-    else:
-        ds_Sv = echopype.calibrate.compute_Sv(echodata)
-        range_output = echodata.compute_range(env_params=[])
-        nan_locs_backscatter_r = ~echodata["Sonar/Beam_group1"].backscatter_r.isel(beam=0).drop("beam").isnull()
-
-    nan_locs_Sv_range = ~ds_Sv.echo_range.isnull()
-    nan_locs_range = ~range_output.isnull()
-    assert xr.Dataset.equals(nan_locs_backscatter_r, nan_locs_range)
-    assert xr.Dataset.equals(nan_locs_backscatter_r, nan_locs_Sv_range)
-
-
-@pytest.mark.parametrize(
-    ["ext_type", "sonar_model", "updated", "path_model", "raw_path", "platform_data"],
-    [
-        (
-                "external-trajectory",
-                "EK80",
-                ("pitch", "roll", "longitude", "latitude"),
-                "EK80",
-                (
-                        "saildrone",
-                        "SD2019_WCS_v05-Phase0-D20190617-T125959-0.raw",
-                ),
-                (
-                        "saildrone",
-                        "saildrone-gen_5-fisheries-acoustics-code-sprint-sd1039-20190617T130000-20190618T125959-1_hz-v1.1595357449818.nc",  #noqa
-                ),
-        ),
-        (
-                "fixed-location",
-                "EK60",
-                ("longitude", "latitude"),
-                "EK60",
-                (
-                        "ooi",
-                        "CE02SHBP-MJ01C-07-ZPLSCB101_OOI-D20191201-T000000.raw"
-                ),
-                (-100.0, -50.0),
-        ),
-    ],
-)
-def test_update_platform(
-        ext_type,
-        sonar_model,
-        updated,
-        path_model,
-        raw_path,
-        platform_data,
-        test_path
-):
-    raw_file = test_path[path_model] / raw_path[0] / raw_path[1]
-    ed = echopype.open_raw(raw_file, sonar_model=sonar_model)
-
-    for variable in updated:
-        assert np.isnan(ed["Platform"][variable].values).all()
-
-    if ext_type == "external-trajectory":
-        extra_platform_data_file_name = platform_data[1]
-        extra_platform_data = xr.open_dataset(
-            test_path[path_model] / platform_data[0] / extra_platform_data_file_name
-        )
-    elif ext_type == "fixed-location":
-        extra_platform_data_file_name = None
-        extra_platform_data = xr.Dataset(
-            {
-                "longitude": (["time"], np.array([float(platform_data[0])])),
-                "latitude": (["time"], np.array([float(platform_data[1])])),
-            },
-            coords={
-                "time": (["time"], np.array([ed['Sonar/Beam_group1'].ping_time.values.min()]))
-            },
-        )
-
-    ed.update_platform(
-        extra_platform_data,
-        extra_platform_data_file_name=extra_platform_data_file_name,
-    )
-
-    for variable in updated:
-        assert not np.isnan(ed["Platform"][variable].values).all()
-
-    # times have max interval of 2s
-    # check times are > min(ed["Sonar/Beam_group1"]["ping_time"]) - 2s
-    assert (
-        ed["Platform"]["time1"]
-        > ed["Sonar/Beam_group1"]["ping_time"].min() - np.timedelta64(2, "s")
-    ).all()
-    # check there is only 1 time < min(ed["Sonar/Beam_group1"]["ping_time"])
-    assert (
-        np.count_nonzero(
-            ed["Platform"]["time1"] < ed["Sonar/Beam_group1"]["ping_time"].min()
-        )
-        <= 1
-    )
-    # check times are < max(ed["Sonar/Beam_group1"]["ping_time"]) + 2s
-    assert (
-        ed["Platform"]["time1"]
-        < ed["Sonar/Beam_group1"]["ping_time"].max() + np.timedelta64(2, "s")
-    ).all()
-    # check there is only 1 time > max(ed["Sonar/Beam_group1"]["ping_time"])
-    assert (
-        np.count_nonzero(
-            ed["Platform"]["time1"] > ed["Sonar/Beam_group1"]["ping_time"].max()
-        )
-        <= 1
-    )
